fix unit test for specific plattforms
this commit should fix the troubles by moving specific files to its places to ensure that the build can be the done. Additionally changing the naming to remove the test to ensure proper importing
