+++ .github/workflows/client-tests.yml
@@ -8,7 +8,9 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [windows-latest, ubuntu-latest]
+        os: [windows-latest, ubuntu-latest, macOS-latest]
+    env:
+      GO111MODULE: off
 
     steps:
 
@@ -17,6 +19,14 @@ jobs:
         with:
           go-version: 1.12
         id: go
+
+      - name: Set GOPATH
+        # temporary fix
+        # see https://github.com/actions/setup-go/issues/14
+        run: |
+          echo "##[set-env name=GOPATH;]$(dirname $GITHUB_WORKSPACE)"
+          echo "##[add-path]$(dirname $GITHUB_WORKSPACE)/bin"
+        shell: bash
 
       - name: Checkout
         uses: actions/checkout@v1
@@ -29,6 +39,12 @@ jobs:
           echo "##[set-env name=CGO_ENABLED;]0"
         shell: bash
 
+      - name: Install deps
+        run: |
+          go get github.com/rogpeppe/godeps
+          godeps -u dependencies.tsv
+
       - name: Test client
         run: |
+          echo "$GOPATH"
           go test -v ./... --check.v
